update sentry and improve logging info
