+++ package.json
@@ -48,8 +48,8 @@
     "@gnosis.pm/safe-apps-sdk": "4.0.0",
     "@metamask/detect-provider": "^1.2.0",
     "@popperjs/core": "^2.9.2",
-    "@sentry/browser": "^6.3.2",
-    "@sentry/tracing": "^6.3.2",
+    "@sentry/browser": "^7.17.4",
+    "@sentry/tracing": "^7.17.4",
     "@sentry/webpack-plugin": "^1.15.1",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/vue": "^6.4.2",
