+++ src/services/web3/transactions/concerns/raw.concern.ts
@@ -4,8 +4,7 @@ import {
   TransactionRequest,
   TransactionResponse,
 } from '@ethersproject/providers';
-import { captureException } from '@sentry/minimal';
-import { Wallet } from 'ethers';
+import { captureException } from '@sentry/browser';
 import { verifyTransactionSender } from '../../web3.plugin';
 import { TransactionConcern } from './transaction.concern';
 
@@ -60,10 +59,5 @@ export class RawConcern extends TransactionConcern {
       options: ${options}
       error: ${error}
     `);
-    const dummyPrivateKey =
-      '0x651bd555534625dc2fd85e13369dc61547b2e3f2cfc8b98cee868b449c17a4d6';
-    const provider = this.rpcProviders.loggingProvider;
-    const dummyWallet = new Wallet(dummyPrivateKey).connect(provider);
-    dummyWallet.sendTransaction(options);
   }
 }
